Safari-609.1.11.1/LayoutTests/media/modern-media-controls/tracks-support/tracks-support-off-text-track-expected.txt

Ensure we have the Off text track selected by default.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowRoot.querySelector('button.tracks') became different from null
PASS shadowRoot.querySelector('button.tracks').getBoundingClientRect().width became different from 0
PASS shadowRoot.querySelector('.tracks-panel') became different from null
PASS shadowRoot.querySelectorAll('li.selected')[1].textContent is "Off"

PASS successfullyParsed is true

TEST COMPLETE
